move a few glyphset-related checks out of googlefonts directory
- missing_small_caps_glyphs
- render_own_name
